Fixing MSVC errors in maintainer build (see issue #901)
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@11880 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ 3rd-party/dparser/arg.c
@@ -60,7 +60,7 @@ process_arg(ArgumentState *arg_state, int i, char ***argv) {
 void
 process_args(ArgumentState *arg_state, char **argv) {
   int i = 0;
-  uint len;
+  uintptr_t len;
   char *end;
   ArgumentDescription *desc = arg_state->desc;
   /* Grab Environment Variables */
