+++ 3rd-party/dparser/read_binary.h
@@ -8,8 +8,8 @@ typedef struct BinaryTablesHead {
   int n_relocs;
   int n_strings;
   int d_parser_tables_loc;
-  int tables_size;
-  int strings_size;
+  size_t tables_size;
+  size_t strings_size;
 } BinaryTablesHead;
 
 typedef struct BinaryTables {
